feat: v3 generation scraper
