+++ packages/scraper/src/gen/marshall.ts
@@ -1,7 +1,10 @@
 import { consola } from "consola";
-import type { Course, Section } from "../types";
 import * as z from "zod";
-import { BannerSection } from "../schemas/section";
+import { BannerSection } from "../schemas/banner/section";
+import {
+  ScraperBannerCacheCourse,
+  ScraperBannerCacheSection,
+} from "../schemas/scraper/banner-cache";
 
 /* arrangeCourses takes the raw sections scraped from banner and
  *
@@ -11,18 +14,39 @@ import { BannerSection } from "../schemas/section";
 export function arrangeCourses(
   bannerSections: z.infer<typeof BannerSection>[],
 ) {
-  const courses: { [key: string]: Course } = {};
+  consola.start("stubbing courses");
+
+  const courses: { [key: string]: z.infer<typeof ScraperBannerCacheCourse> } =
+    {};
   const xlist: { [key: string]: string[] } = {};
-  const sections: { [key: string]: Section[] } = {};
-  const subjects: string[] = [];
-  const campuses: string[] = [];
-  const buildings = [];
+  const sections: {
+    [key: string]: z.infer<typeof ScraperBannerCacheSection>[];
+  } = {};
+  const subjectCodes: string[] = [];
+  const attributes: Map<string, string> = new Map();
+  const campuses: Map<string, { code: string; description: string }> =
+    new Map();
+  const buildings: Map<
+    string,
+    { code: string; description: string; campus: string }
+  > = new Map();
+  const rooms: Map<
+    string,
+    { rooms: Set<string>; building: string; campus: string }
+  > = new Map();
 
   const crns: string[] = [];
 
   for (const s of bannerSections) {
     if (!Object.keys(courses).includes(s.subjectCourse)) {
       const specialTopics = s.courseTitle.includes("Special Topics");
+
+      const courseAttributes = s.sectionAttributes.map((a) => ({
+        code: a.code,
+        description: a.description.trim(),
+      }));
+
+      courseAttributes.forEach((n) => attributes.set(n.code, n.description));
 
       courses[s.subjectCourse] = {
         // special topic courses are when course information is section scoped
@@ -34,10 +58,7 @@ export function arrangeCourses(
         description: "", // note - this will be filled in later when the descriptions are scraped
         maxCredits: s.creditHourHigh ?? s.creditHourLow,
         minCredits: s.creditHourLow,
-        // Store raw nupath codes from banner - will be filtered during upload
-        nupath: s.sectionAttributes
-          .filter((a) => a.code.startsWith("NC"))
-          .map((a) => a.code.replace("NC", "")),
+        attributes: courseAttributes.map((n) => n.code),
         prereqs: {},
         coreqs: {},
         postreqs: {},
@@ -48,8 +69,9 @@ export function arrangeCourses(
 
     if (
       !c.specialTopics &&
-      (s.courseTitle !== c.name || // special topic courses have different names between sections
-        s.sectionAttributes.filter((s) => s.code === "TOPC").length > 0) // or sometimes have a "Topics" attribute
+      // (s.courseTitle !== c.name || // special topic courses have different names between sections
+      //   s.sectionAttributes.filter((s) => s.code === "TOPC").length > 0) // or sometimes have a "Topics" attribute
+      s.sectionAttributes.filter((s) => s.code === "TOPC").length > 0
     ) {
       c.specialTopics = true;
       c.name = "Special Topics"; // update name placeholder (special topic course names are scraped later)
@@ -67,6 +89,12 @@ export function arrangeCourses(
 
     crns.push(s.courseReferenceNumber);
 
+    const {
+      meetingTimes,
+      campuses: mtCampuses,
+      buildings: mtBuildings,
+    } = parseMeetingTimes(s);
+
     sections[s.subjectCourse].push({
       crn: s.courseReferenceNumber,
       name: s.courseTitle,
@@ -78,9 +106,8 @@ export function arrangeCourses(
       waitlistRemaining: s.waitAvailable,
       classType: s.scheduleTypeDescription,
       honors: s.sectionAttributes.some((a) => a.description === "Honors"),
-      // Store raw campus code from banner - will be validated during upload
       campus: s.campusDescription,
-      meetingTimes: parseMeetingTimes(s),
+      meetingTimes: meetingTimes,
 
       faculty: [],
 
@@ -90,12 +117,69 @@ export function arrangeCourses(
       xlist: s.crossList ? xlist[s.crossList] : [],
     });
 
-    if (!subjects.includes(s.subject)) subjects.push(s.subject);
-    if (!campuses.includes(s.campusDescription))
-      campuses.push(s.campusDescription);
+    if (!subjectCodes.includes(s.subject)) subjectCodes.push(s.subject);
+
+    // if (mtCampuses.length === 0 && meetingTimes.length > 0) {
+    //   consola.warn("no campuses across meeting times", {
+    //     crn: s.courseReferenceNumber,
+    //     sectionCampus: s.campusDescription,
+    //     mt: mtCampuses,
+    //   });
+    // }
+
+    if (mtCampuses.length > 0) {
+      campuses.set(mtCampuses[0].description, {
+        code: mtCampuses[0].code,
+        description: mtCampuses[0].description,
+      });
+    } else if (!campuses.has(s.campusDescription)) {
+      campuses.set(s.campusDescription, {
+        code: "?",
+        description: s.campusDescription,
+      });
+    }
+
+    mtBuildings.forEach((v, k) => {
+      if (mtCampuses.length === 0) {
+        consola.warn("no campus specified for a set building", {
+          crn: s.courseReferenceNumber,
+        });
+        return;
+      }
+
+      const campusCode = mtCampuses[0].code;
+      if (!buildings.has(k)) {
+        buildings.set(k, {
+          code: k,
+          description: v.description,
+          campus: campusCode,
+        });
+      }
+
+      if (!rooms.has(k)) {
+        rooms.set(k, { rooms: new Set(), building: k, campus: campusCode });
+      }
+      const roomRef = rooms.get(k);
+      if (roomRef) {
+        v.rooms.forEach((r) => roomRef.rooms.add(r));
+      }
+    });
   }
 
-  return { courses: Object.values(courses), sections, subjects, campuses };
+  // consola.box("campuses", campuses);
+  // consola.box("attributes", attributes);
+  // consola.box("buildings", buildings);
+  // consola.box("rooms", rooms);
+
+  return {
+    courses: Object.values(courses),
+    sections,
+    subjects: subjectCodes,
+    attributes,
+    campuses,
+    buildings,
+    rooms,
+  };
 }
 
 /*
@@ -107,11 +191,19 @@ export function arrangeCourses(
 export function parseMeetingTimes(section: z.infer<typeof BannerSection>) {
   // BUG: somewhere in here lol
   const meetings = [];
+  const buildings: Map<string, { description: string; rooms: string[] }> =
+    new Map();
+  const campuses: { code: string; description: string }[] = [];
+
   for (const meetingFaculty of section.meetingsFaculty) {
-    const meetingTime = meetingFaculty?.meetingTime;
-    if (!meetingTime) {
-      consola.info("no meeting time " + section.courseReferenceNumber);
-      continue;
+    const meetingTime = meetingFaculty.meetingTime;
+
+    //
+    if (meetingTime.campus && meetingTime.campusDescription) {
+      campuses.push({
+        code: meetingTime.campus,
+        description: meetingTime.campusDescription,
+      });
     }
 
     const days = [];
@@ -136,20 +228,34 @@ export function parseMeetingTimes(section: z.infer<typeof BannerSection>) {
       meetingTime.meetingType === "FNEX";
 
     // Only include finalDate if it's a final exam
-    let finalDate = null;
+    let finalDate: string | null = null;
     if (isFinal && meetingTime.startDate) {
       finalDate = meetingTime.startDate;
     }
 
-    let actualRoom;
-    if (!meetingTime.room) {
-      actualRoom = "";
+    let actualRoom: string | null = null;
+    if (!meetingTime.room || meetingTime.room === "ROOM") {
+      actualRoom = null;
     } else {
-      actualRoom = meetingTime.room == "ROOM" ? "" : meetingTime.room;
+      actualRoom = meetingTime.room;
+    }
+
+    if (meetingTime.building && meetingTime.buildingDescription) {
+      if (!buildings.has(meetingTime.building)) {
+        buildings.set(meetingTime.building, {
+          description: meetingTime.buildingDescription,
+          rooms: [],
+        });
+      }
+
+      const b = buildings.get(meetingTime.building);
+      if (b && actualRoom) {
+        b.rooms.push(actualRoom);
+      }
     }
 
     meetings.push({
-      building: meetingTime.buildingDescription || meetingTime.building || "",
+      building: meetingTime.building,
       room: actualRoom,
       days: days,
       startTime: startTime,
@@ -159,5 +265,5 @@ export function parseMeetingTimes(section: z.infer<typeof BannerSection>) {
     });
   }
 
-  return meetings;
+  return { meetingTimes: meetings, campuses, buildings };
 }
