+++ packages/scraper/src/gen/pieces/courseNames.ts
@@ -3,7 +3,7 @@ import { FetchEngine } from "../fetch";
 import { sectionCatalogDetailsEndpoint } from "../endpoints";
 import { consola } from "consola";
 import { decode } from "html-entities";
-import { BannerSectionCatalogDetails } from "../../schemas/sectionCatalogDetails";
+import { BannerSectionCatalogDetails } from "../../schemas/banner/sectionCatalogDetails";
 
 /**
  *
@@ -59,10 +59,6 @@ export async function scrapeCatalogDetails(
       }
 
       const catalogDetails = catalogDetailsResult.data;
-
-      if (!catalogDetails) {
-        return;
-      }
 
       c.name =
         decode(decode(catalogDetails))
