+++ packages/scraper/src/schemas/banner/common.ts
@@ -1,0 +1,4 @@
+import * as z from "zod";
+
+export const BannerTerm = z.string().length(6);
+export const BannerCRN = z.string().length(5);
