+++ packages/scraper/src/schemas/banner/sectionCoreqs.ts
@@ -1,0 +1,3 @@
+import * as z from "zod";
+
+export const BannerSectionCoreqs = z.string();
