+++ packages/scraper/src/schemas/banner/terms.ts
@@ -1,0 +1,9 @@
+import * as z from "zod";
+import { BannerTerm } from "./common";
+
+export const BannerTerms = z.strictObject({
+  code: BannerTerm,
+  description: z.string(),
+});
+
+export const BannerTermsResponse = z.array(BannerTerms);
