--- packages/scraper/src/schemas/sectionCatalogDetails.ts
@@ -1,3 +1,0 @@
-import * as z from "zod";
-
-export const BannerSectionCatalogDetails = z.string();
