--- packages/scraper/src/schemas/sectionCoreqs.ts
@@ -1,3 +1,0 @@
-import * as z from "zod";
-
-export const BannerSectionCoreqs = z.string();
