--- packages/scraper/src/schemas/sectionDescription.ts
@@ -1,3 +1,0 @@
-import * as z from "zod";
-
-export const BannerSectionDescription = z.string();
