--- packages/scraper/src/schemas/sectionPrereqs.ts
@@ -1,3 +1,0 @@
-import * as z from "zod";
-
-export const BannerSectionPrereqs = z.string();
