+++ packages/scraper/src/scripts/generate.ts
@@ -1,4 +1,3 @@
-import { TermScrape } from "../types";
 import { existsSync, readFileSync, writeFileSync } from "node:fs";
 import path from "node:path";
 import { parse } from "yaml";
@@ -7,13 +6,11 @@ import { scrapeCatalogTerm } from "../gen/main";
 import { infer as zinfer } from "zod";
 import { Config } from "../config";
 import { consola } from "consola";
+import { ScraperBannerCache } from "../schemas/scraper/banner-cache";
 
 const CACHE_PATH = "cache/";
 const CACHE_FORMAT = (term: string) => `term-${term}.json`;
-
-interface CachedTermScrape extends TermScrape {
-  timestamp: string;
-}
+const CACHE_VERSION = 3;
 
 const main = defineCommand({
   meta: {
@@ -21,20 +18,43 @@ const main = defineCommand({
     description: "runs the scraper to generate the banner cache files",
   },
   args: {
-    cachePath: {
-      type: "positional",
-      description: "path of the cache files",
-      required: false,
-    },
     terms: {
       type: "string",
+      default: "active",
       description:
-        "comma-seperated list of specific terms to scrape, 'active', or 'all'",
+        "comma-seperated list of specific terms to scrape, 'active' (default), or 'all'",
+      required: false,
+    },
+    interactive: {
+      alias: "i",
+      type: "boolean",
+      description: "",
+      required: false,
+    },
+    overwrite: {
+      alias: "f",
+      type: "boolean",
+      description: "overwrites existing caches rather than skipping",
+      required: false,
+    },
+    verbose: {
+      alias: "v",
+      type: "boolean",
+      description: "",
+      required: false,
+    },
+    veryverbose: {
+      alias: "vv",
+      type: "boolean",
+      description: "",
       required: false,
     },
   },
   async run({ args }) {
-    consola.level = 999;
+    if (args.verbose) consola.level = 4;
+    if (args.veryverbose) consola.level = 999;
+
+    const interactive = args.interactive;
 
     const configStream = readFileSync(
       path.resolve(CACHE_PATH, "manifest.yaml"),
@@ -45,39 +65,50 @@ const main = defineCommand({
     const configRaw = parse(configStream);
     const configResponse = Config.safeParse(configRaw);
     if (!configResponse.success) {
-      console.error(configResponse.error);
+      consola.error(configResponse.error);
       return;
     }
 
     const config = configResponse.data;
 
-    const termsToScrape = filterTerms(config, { terms: [], all: true });
+    const termsToScrape = filterTerms(config, args.terms);
+    consola.info(`scraping ${termsToScrape.length} terms`);
 
     if (termsToScrape.length === 0) {
-      console.log("ℹ️  No active terms to scrape");
+      consola.log("no active / configured terms to scrape");
       return;
     }
 
     for (const termConfig of termsToScrape) {
-      if (termConfig.term !== 202630) continue;
+      consola.start(`scraping term ${termConfig.term}`);
+
       const cachename = path.resolve(
         CACHE_PATH,
         CACHE_FORMAT(termConfig.term.toString()),
       );
       const existingCache = existsSync(cachename);
+      if (args.overwrite && existingCache) {
+        consola.info("existing cache found, overwriting with new scrape");
+      } else if (!args.overwrite && existingCache) {
+        consola.success("existing cache found, skipping term");
+        continue;
+      }
 
       const out = await scrapeCatalogTerm(
         termConfig.term.toString(),
         termConfig,
+        interactive,
       );
 
       if (!out) {
+        consola.error(`error scraping term ${termConfig.term}`);
         return;
       }
 
-      const cachedData: CachedTermScrape = {
-        ...out,
+      const cachedData: zinfer<typeof ScraperBannerCache> = {
+        version: CACHE_VERSION,
         timestamp: new Date().toISOString(),
+        ...out,
       };
 
       writeFileSync(cachename, JSON.stringify(cachedData, null, 2));
@@ -85,36 +116,32 @@ const main = defineCommand({
     }
 
     consola.success(
-      `🎉 Successfully scraped ${termsToScrape.length} term${termsToScrape.length > 1 ? "s" : ""}`,
+      `successfully scraped ${termsToScrape.length} term${termsToScrape.length > 1 ? "s" : ""}`,
     );
   },
 });
 
 void runMain(main);
 
-function filterTerms(
-  config: zinfer<typeof Config>,
-  options: { terms: string[]; all: boolean },
-) {
-  const now = new Date();
-
-  if (options.terms.length > 0) {
-    const filteredTerms = config.terms.filter((t) =>
-      options.terms.includes(t.term.toString()),
-    );
-    if (filteredTerms.length === 0) {
-      console.log(
-        `⚠️  No matching terms found for: ${options.terms.join(", ")}`,
-      );
-      process.exit(1);
-    }
-    return filteredTerms;
-  }
-
-  if (options.all) {
+/**
+ */
+function filterTerms(config: zinfer<typeof Config>, termArg: string) {
+  if (termArg === "all") {
     return config.terms;
   }
 
-  // Default: only active terms
-  return config.terms.filter((t) => new Date(t.activeUntil) > now);
+  if (termArg === "active") {
+    const now = new Date();
+    return config.terms.filter((t) => new Date(t.activeUntil) > now);
+  }
+
+  const splitTerms = termArg.split(",");
+  const filteredTerms = config.terms.filter((t) =>
+    splitTerms.includes(t.term.toString()),
+  );
+  if (filteredTerms.length === 0) {
+    consola.error(`no matching terms found for: ${splitTerms.join(", ")}`);
+    process.exit(1);
+  }
+  return filteredTerms;
 }
